+++ HISTORY.rst
@@ -3,6 +3,13 @@
 History
 =======
 
+2.0.3 (unreleased)
+------------------
+
+This is a bugfix-only version:
+
+- When processing stripe event, wrap create object call in transaction (#877)
+
 2.0.2 (2019-06-09)
 ------------------